fix(clash): recheck the abort signal on the way back from the yield
Review follow-ups on the cancellation path.

The yield IS the window an abort arrives in — a deadline timer or a UI
handler runs during that turn of the event loop and nowhere else — but
the signal was only checked on the way in, so a run cancelled during its
own yield spent another 256 pairs before noticing. Rechecked after the
await now, before `lastYield` is updated or another candidate is
processed.

Two ordering fixes alongside it: an already-aborted signal throws before
the broad phase rather than after it, and the periodic checkpoint runs
ahead of the `maxCandidatePairs` exit so a cancelled run reports the
cancellation instead of a quietly truncated result (`maxPairs === 0`
broke out before ever looking at the signal).

`ClashSettings.signal` now documents its real bound rather than
overstating it: the check is every 256 pairs, but the *yield* only
happens past the yield interval, so a run that finishes inside that
first interval never returns to the event loop and cannot observe an
abort raised after it began.

Test determinism, which was the other half of the load-sensitivity fixed
in the previous commit: `TsKernel` takes its yield interval as a
constructor argument, so the cancellation tests drive the shipping loop
at a fixed *pair* cadence instead of racing the host's clock. They no
longer depend on how fast the machine is, and the suite runs in ~300ms.

Also: cancellation coverage for the real `ClashNamespace.matrix`, whose
options destructure is one added name away from silently dropping the
signal for the whole discipline matrix; a disposed `HostWorkQueue` no
longer installs a live controller when `settleHostWork` abandons on its
way out; and the sandbox cancellation suite moves to its own file
(bridge-async.test.ts is back to its pre-PR 445 lines).

Measured, interleaved A/B over 14,400 real candidate pairs, 10 runs per
arm: the 256-pair cadence is indistinguishable from 1024 (medians 880.0
vs 880.1 ms unsignalled, 901.5 vs 900.2 ms signalled, identical clash
counts). Supplying a signal at all costs ~2.4%, from the yields, and the
uncancellable broad phase is 0.16% of that run.
