+++ packages/clash/src/engine-ts/ts-kernel.ts
@@ -16,6 +16,14 @@ import type { ClashKernel, NarrowRecord, RuleDetection } from './kernel.js';
 export class TsKernel implements ClashKernel {
   private readonly triCache = new WeakMap<ClashElement, TriMesh>();
 
+  /**
+   * @param yieldMs How long the narrow phase may hold the thread between yields.
+   * The default is a few frames, which is what keeps a main-thread run painting.
+   * A caller that needs the loop to reach the event loop on a fixed cadence
+   * rather than a time-based one (the cancellation tests do) passes `0`.
+   */
+  constructor(private readonly yieldMs: number = YIELD_MS) {}
+
   prepare(): void {
     // Triangle BVHs are built lazily per element on first use, and cached for
     // the lifetime of this kernel so an element shared across rules pays once.
@@ -46,6 +54,12 @@ export class TsKernel implements ClashKernel {
     const resolveBIdx = groupBIdx ?? groupAIdx;
     const margin = Math.max(tolerance, rule.clearance ?? 0);
 
+    // Before the broad phase, not after it: a signal that is already aborted
+    // when the rule starts must not buy a full candidate-pair build first.
+    // Redundant with the checkpoint at the loop's first iteration, so its only
+    // signature is work not done, not a different outcome.
+    if (signal?.aborted) throw abortError();
+
     const pairs = candidatePairs(groupA, groupB, margin);
     const total = pairs.length;
     const records: NarrowRecord[] = [];
@@ -65,10 +79,6 @@ export class TsKernel implements ClashKernel {
     const canInterrupt = onProgress !== undefined || signal !== undefined;
 
     for (const [i, j] of pairs) {
-      if (processed >= maxPairs) {
-        candidatesDropped = total - processed;
-        break;
-      }
       // Every 256 pairs: check cancellation, and if we've held the thread for
       // more than a frame's worth of time, report progress and yield so the UI
       // can repaint and stay responsive on large models.
@@ -78,16 +88,30 @@ export class TsKernel implements ClashKernel {
       // between two real building elements is not cheap — 1024 of them is a
       // visible stretch of CPU to spend after the caller has given up. The
       // check itself is a property read plus a clock read against ~256 BVH
-      // traversals, so the finer cadence costs nothing measurable.
+      // traversals, so the finer cadence costs nothing measurable (measured
+      // interleaved A/B over 14,400 real pairs: within run-to-run noise).
+      //
+      // Ahead of the `maxPairs` exit below, so a cancelled run reports the
+      // cancellation rather than a quietly truncated result — including the
+      // degenerate `maxPairs === 0`, where the loop would otherwise break out
+      // before ever looking at the signal.
       if ((processed & 0xff) === 0) {
-        if (signal?.aborted) {
-          throw new DOMException('Clash run aborted', 'AbortError');
-        }
-        if (canInterrupt && now() - lastYield > YIELD_MS) {
+        if (signal?.aborted) throw abortError();
+        if (canInterrupt && now() - lastYield > this.yieldMs) {
           onProgress?.(processed, total);
           await yieldToEventLoop();
+          // Rechecked after the await, because the yield IS the window the
+          // abort arrives in: a deadline timer or a UI handler runs during this
+          // turn of the event loop and nowhere else. Checking only before it
+          // would spend another 256 pairs on a run that had already been
+          // cancelled, which is the waste this whole path exists to remove.
+          if (signal?.aborted) throw abortError();
           lastYield = now();
         }
+      }
+      if (processed >= maxPairs) {
+        candidatesDropped = total - processed;
+        break;
       }
       processed += 1;
       const elA = groupA[i];
@@ -111,6 +135,11 @@ export class TsKernel implements ClashKernel {
 
 /** Hold the main thread no longer than this between yields (≈ a few frames). */
 const YIELD_MS = 50;
+
+/** The rejection a cancelled run produces, spelled the same at every checkpoint. */
+function abortError(): DOMException {
+  return new DOMException('Clash run aborted', 'AbortError');
+}
 
 function now(): number {
   return typeof performance !== 'undefined' ? performance.now() : Date.now();
